refactor: remove util file
